# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tracepad API",
    "summary": "Lightweight, self-hosted, OTLP-native store for LLM and agent traces.",
    "description": "The read API is the product's contract: the UI, the CLI and the MCP server are thin clients over exactly these endpoints, and a feature that is not here does not exist. Responses are flat snake_case JSON with RFC 3339 timestamps and opaque cursors. Every response spends a byte budget on payloads (default 51200, `?budget=` to override): a payload that does not fit is cut at a UTF-8 boundary and replaced by a truncation marker naming where the whole of it lives.",
    "license": {"name": "Apache-2.0", "identifier": "Apache-2.0"},
    "version": "0.1.0"
  },
  "servers": [{"url": "http://localhost:4318", "description": "A locally running tracepad"}],
  "security": [{"bearerKey": []}, {"basicKeys": []}],
...
  "paths": {
...
    "/api/v1/system": {
      "get": {
        "tags": ["meta"],
        "summary": "Version, uptime, database size and ingest counters since start",
        "description": "Self-diagnosis without a metrics stack. The counters are in-memory and cover this process only, which is what `counters.since` says. Row counts and counters are the asking project's own — `payloads` is absent because it cannot be attributed to one, `projects` is how many tenants share this process (naming none of them), and `size_bytes` is the whole deployment's file on disk.",
        "operationId": "system",
        "responses": {
          "200": {"description": "What this process knows about itself", "content": {"application/json": {"schema": {
            "type": "object",
            "properties": {
              "version": {"type": "string"},
              "go_version": {"type": "string"},
              "started_at": {"type": "string", "format": "date-time"},
              "uptime_seconds": {"type": "integer"},
              "database": {"type": "object", "properties": {
                "size_bytes": {"type": "integer", "description": "The database file plus its write-ahead log, for the whole deployment"},
                "rows": {"type": "object", "additionalProperties": {"type": "integer"},
                  "description": "Row counts within the asking project, plus `projects` as a bare tenant count"}
              }},
              "writer_queue": {"type": "object", "properties": {
                "waiting": {"type": "integer"}, "capacity": {"type": "integer"}
              }},
              "read_slots": {"type": "object",
                "description": "Reads being served now, for the whole deployment, against TRACEPAD_READ_CONCURRENCY; a read that finds none free waits within its deadline",
                "properties": {
                  "busy": {"type": "integer"}, "capacity": {"type": "integer"}
                },
                "required": ["busy", "capacity"]},
              "response_budget_bytes": {"type": "integer"},
              "runs": {"type": "object",
                "description": "The link between traces and dataset runs, as this project sees it",
                "properties": {
                  "pinned_traces": {"type": "integer", "description": "Traces a live run keeps out of the retention sweep — the size of retention's one exception"},
                  "orphan_traces": {"type": "integer", "description": "Trace deliveries since start that named a run this project does not have; they are stored, and each unknown id is logged once"}
                },
                "required": ["pinned_traces", "orphan_traces"]},
              "raw": {"type": "object",
                "description": "The archive of export bodies: what an export can replay, and what it cannot. Unlike the counters this is on disk, not since start.",
                "properties": {
                  "enabled": {"type": "boolean", "description": "Whether accepted bodies are still being kept (TRACEPAD_STORE_RAW)"},
                  "batches": {"type": "integer"},
                  "bytes": {"type": "integer", "description": "What the archive occupies on disk, compressed — the number an operator moving `raw_retention_days` is deciding about"},
                  "oldest_received_at": {"type": ["string", "null"], "format": "date-time"},
                  "newest_received_at": {"type": ["string", "null"], "format": "date-time"},
                  "traces_before_window": {"type": "integer",
                    "description": "Traces whose earliest span started before the oldest batch arrived, and which an export therefore cannot cover. A lower bound: a late export of an old trace lands after the batch line."}
                },
                "required": ["enabled", "batches", "bytes", "oldest_received_at", "newest_received_at", "traces_before_window"]},
              "media": {"type": "object",
                "description": "The images and files ingest took out of this project's payloads: the setting that decides whether they are kept, and what the kept ones cost. A body two projects share is in both projects' figures.",
                "properties": {
                  "setting": {"type": "string", "enum": ["store", "placeholder"]},
                  "count": {"type": "integer", "description": "Distinct bodies this project's traces and raw batches point at"},
                  "bytes": {"type": "integer", "description": "Their decoded size"}
                },
                "required": ["setting", "count", "bytes"]},
              "compaction": {"type": "object",
                "description": "The deployment's compaction: after an erasure, a trace deletion or a project's purge, the next sweeper pass merges the search index, drains the freelist and truncates the write-ahead log, so what the deletion unlinked is overwritten rather than left in the file",
                "properties": {
                  "requested_at": {"type": ["string", "null"], "format": "date-time", "description": "The latest request still pending; null when none is"},
                  "completed_at": {"type": ["string", "null"], "format": "date-time", "description": "When one last finished; null before the first"}
                },
                "required": ["requested_at", "completed_at"]},
              "counters": {"type": "object",
                "description": "This project's ingest traffic, and its reads the read bounds refused, since the process started",
                "properties": {
                "since": {"type": "string", "format": "date-time"},
                "dialects": {"type": "object", "additionalProperties": {"type": "object", "properties": {
                  "batches": {"type": "integer"},
                  "spans_stored": {"type": "integer"},
                  "spans_skipped": {"type": "integer"}
                }}},
                "rejected_batches": {"type": "integer"},
                "unreadable_resource_spans": {"type": "integer"},
                "langfuse_ingestion_versions": {"type": "array", "items": {"type": "string"}},
                "reads_timed_out": {"type": "integer", "description": "This project's reads the read deadline stopped (TRACEPAD_READ_TIMEOUT)"},
                "reads_refused_busy": {"type": "integer", "description": "This project's reads refused because no read slot came free before the deadline"}
              }}
            },
            "required": ["version", "started_at", "uptime_seconds", "database", "counters"]
          }}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
...
    "/api/v1/raw/{id}": {
      "get": {
        "tags": ["raw"],
        "summary": "One raw export body, in the Content-Type it was received in",
        "description": "The bytes the client posted, gzip already removed. Exempt from the response budget for the reason `/observations/{id}/io` is: a cut body is not a smaller batch, it is a broken one. `X-Tracepad-Received-At` and `X-Tracepad-Dialect` carry what a client needs to name a file or label a manifest without a second request.",
        "operationId": "rawBatch",
        "parameters": [
          {"name": "id", "in": "path", "required": true,
            "schema": {"type": "integer", "minimum": 1}}
        ],
        "responses": {
          "200": {"description": "The body as received",
            "headers": {
              "X-Tracepad-Received-At": {"schema": {"type": "string", "format": "date-time"}},
              "X-Tracepad-Dialect": {"schema": {"type": "string"},
                "description": "Absent when the mapper claimed nothing"}
            },
            "content": {
              "application/x-protobuf": {"schema": {"type": "string", "format": "binary"}},
              "application/json": {"schema": {"type": "string", "format": "binary"}}
            }},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerKey": {"type": "http", "scheme": "bearer",
        "description": "A project's secret key, `tp-sk-…`"},
      "basicKeys": {"type": "http", "scheme": "basic",
        "description": "`base64(public key:secret key)`, the shape a Langfuse SDK sends"},
      "adminToken": {"type": "http", "scheme": "bearer",
        "description": "`TRACEPAD_ADMIN_TOKEN`, sent in the same header as a project key. Required for everything cross-project, for creating and deleting a project, and for the account routes; beside an owner's or editor's session, it is what lists, mints and revokes a project's keys, which no project key does — which is what makes it the recovery when every owner's password is lost. A server with none configured answers 403 and says so."},
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "tracepad_session",
        "description": "A browser session: `HttpOnly`, `SameSite=Lax`, `Secure` over TLS, set by the three ways in and lasting `TRACEPAD_SESSION_DAYS` (default 30), sliding. On a route whose policy is `member` or `editor` a session must also name the project it is asking about with `X-Tracepad-Project`, and an unsafe method must carry an `Origin` (or `Referer`) whose host is the request's `Host`, the first value of `X-Forwarded-Host`, or the host of `TRACEPAD_URL`."}
    },
...
    "responses": {
      "BadRequest": {"description": "The request cannot mean what it says: an unknown query parameter, a malformed value, a limit out of range",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "Unauthorized": {"description": "The credentials do not resolve to a project",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "ServiceUnavailable": {"description": "The credentials could not be checked right now, more passwords were being checked at once than the server allows, or a write met a database condition that passes (a full disk, a lock that did not clear), or a read found no free read slot before its deadline (`the server is busy; retry shortly`): retry after `Retry-After`. A read the deadline stopped while it ran (`the read took longer than 20s and was stopped; narrow the time range or the filters`, the number being `TRACEPAD_READ_TIMEOUT`) carries no `Retry-After`: the same request would be stopped again, so narrow it",
        "headers": {"Retry-After": {"schema": {"type": "string"}, "description": "Seconds to wait"}},
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "NotFound": {"description": "No such thing in this project",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The one error shape the whole surface uses",
        "properties": {"error": {"type": "string"}},
        "required": ["error"]
      },
...
    }
  }
}